// app/[lang]/components/Services/types.tsx
// app/[lang]/services/types.tsx

export interface ServicesProps {
    dictionary: {
      title: string;
      frontendTitle: string;
      frontendDescription: string;
      frontendItem1: string;
      frontendItem2: string;
      frontendItem3: string;
      frontendItem4: string;
      aiTitle: string;
      aiDescription: string;
      aiItem1: string;
      aiItem2: string;
      aiItem3: string;
      aiItem4: string;
      readyTitle: string;
      readyText: string;
    };
  }
  